move npu into backends and move python into tools
