+++ lite/api/cxx_api.cc
@@ -19,7 +19,7 @@
 #include <vector>
 #include "lite/utils/io.h"
 #ifdef LITE_WITH_NPU
-#include "lite/npu/npu_helper.h"
+#include "lite/backends/npu/npu_helper.h"
 #endif
 
 namespace paddle {
