+++ lite/backends/npu/bridge/mul_op.cc
@@ -19,9 +19,9 @@
 #include "ai_ddk_lib/include/graph/op/all_ops.h"
 #include "ai_ddk_lib/include/graph/operator.h"
 #include "ai_ddk_lib/include/graph/operator_reg.h"
-#include "lite/npu/bridge/registry.h"
-#include "lite/npu/bridge/utils.h"
-#include "lite/npu/npu_helper.h"
+#include "lite/backends/npu/bridge/registry.h"
+#include "lite/backends/npu/bridge/utils.h"
+#include "lite/backends/npu/npu_helper.h"
 
 namespace paddle {
 namespace lite {
