+++ lite/core/mir/subgraph/generate_npu_program_pass.cc
@@ -27,10 +27,10 @@
 #include "ai_ddk_lib/include/graph/model.h"
 #include "ai_ddk_lib/include/graph/op/all_ops.h"  // for ge::op::Data
 #include "ai_ddk_lib/include/graph/operator_reg.h"
-#include "lite/npu/bridge/paddle_use_npu_bridges.h"
-#include "lite/npu/bridge/registry.h"
-#include "lite/npu/bridge/utils.h"
-#include "lite/npu/npu_helper.h"
+#include "lite/backends/npu/bridge/paddle_use_npu_bridges.h"
+#include "lite/backends/npu/bridge/registry.h"
+#include "lite/backends/npu/bridge/utils.h"
+#include "lite/backends/npu/npu_helper.h"
 
 namespace paddle {
 namespace lite {
